+++ R/01-helperFunctions.R
@@ -107,17 +107,22 @@ updateMatrixNamesInput <- function(session, inputId, value, value2) {
 #' 
 #' @export
 setVarsForUncMatrix <- function(timeVars, indVar, renewalRates, renewalRatesUnc = NULL) {
-  if (is.null(renewalRatesUnc)) {
+  if (length(renewalRatesUnc) == 0 || nrow(renewalRates) != nrow(renewalRatesUnc)) {
+    if (length(renewalRatesUnc) > 0 && nrow(renewalRatesUnc) > 1) {
+      warning("The number of rows differs between renewal rates uncertainty and renewal rates. Uncertainty will be ignored!")
+    }
     notIndOrTime <- colnames(renewalRates)[!colnames(renewalRates) %in% c(timeVars, indVar)]
     renewalRatesUnc <- data.frame(renewalRates)
     renewalRatesUnc[notIndOrTime] <- 0
+    rownamesUnc <- rownames(renewalRates)
   } else {
+    rownamesUnc <- rownames(renewalRatesUnc)
     renewalRatesUnc <- data.frame(renewalRatesUnc)
-    if ((!is.null(timeVars) && all(timeVars != "")) || (!is.null(indVar) && indVar != ""))
+    if ((!is.null(timeVars) && all(timeVars != "")) || (length(indVar) != 0 && indVar != ""))
     renewalRatesUnc[c(timeVars, indVar)] <- data.frame(renewalRates)[c(timeVars, indVar)]
   }
   
   renewalRatesUnc <- as.matrix(renewalRatesUnc)
-  rownames(renewalRatesUnc) <- rep("", nrow(renewalRatesUnc))
+  rownames(renewalRatesUnc) <- rownamesUnc
   renewalRatesUnc
 }