+++ R/03-module_modelSpecifications.R
@@ -35,10 +35,15 @@ modelSpecificationsUI <- function(id, title) {
       multiple = TRUE
     ),
     tags$br(),
-    selectizeInput(
-      inputId = ns("indVar"),
-      label = "Individual variable",
-      choices = character(0)
+    checkboxInput(ns("rownamesAsIndVar"), label = "Use rownames as individual variable", value = FALSE),
+    conditionalPanel(
+      ns = ns,
+      condition = "input.rownamesAsIndVar == false",
+      selectizeInput(
+        inputId = ns("indVar"),
+        label = "Individual variable:",
+        choices = character(0)
+      ),
     ),
     tags$br(),
     sliderInput(inputId = ns("iter"),
@@ -50,6 +55,13 @@ modelSpecificationsUI <- function(id, title) {
     sliderInput(inputId = ns("chains"),
                 label = "MCMC chains:",
                 min = 1, max = 12, step = 1, value = defaultModelSpecValues()$chains),
+    checkboxInput(ns("rndmSeed"), label = "Random seed", value = defaultModelSpecValues()$rndmSeed),
+    conditionalPanel(
+      condition = "input.rndmSeed == false",
+      ns = ns,
+      numericInput(ns("fixedSeed"), label = "Fixed seed value", value = defaultModelSpecValues()$fixedSeed),
+      tags$br()
+    ),
     tags$br()
   )
 }
@@ -84,6 +96,8 @@ modelSpecificationsServer <- function(id, dataMatrix, uploadedModelSpecInputs = 
           updateSliderInput(session, "iter", value = defaultModelSpecValues()$iter)
           updateSliderInput(session, "burnin", value = defaultModelSpecValues()$burnin)
           updateSliderInput(session, "chains", value = defaultModelSpecValues()$chains)
+          updateCheckboxInput(session, "rndmSeed", value = defaultModelSpecValues()$rndmSeed)
+          updateNumericInput(session, "fixedSeed", value = defaultModelSpecValues()$fixedSeed)
         }
         
         req(uploadedModelSpecInputs())
@@ -96,11 +110,15 @@ modelSpecificationsServer <- function(id, dataMatrix, uploadedModelSpecInputs = 
         updateSliderInput(session, "iter", value = uploadedModelSpecInputs()$iter)
         updateSliderInput(session, "burnin", value = uploadedModelSpecInputs()$burnin)
         updateSliderInput(session, "chains", value = uploadedModelSpecInputs()$chains)
+        updateCheckboxInput(session, "rndmSeed", value = uploadedModelSpecInputs()$rndmSeed)
+        updateNumericInput(session, "fixedSeed", value = uploadedModelSpecInputs()$fixedSeed)
       })
       
       observeEvent(input$timeVars, {
-        values$timeMinimum <- dataMatrix()[, input$timeVars[1]] %>% min
-        values$timeMaximum <- dataMatrix()[, input$timeVars[1]] %>% max
+        values$timeMinimum <- getTimeMin(mtrx = dataMatrix(),
+                                         timeVars = input$timeVars)
+        values$timeMaximum <- getTimeMax(mtrx = dataMatrix(),
+                                         timeVars = input$timeVars)
         values$timeVars <- input$timeVars
       })
       
@@ -110,6 +128,15 @@ modelSpecificationsServer <- function(id, dataMatrix, uploadedModelSpecInputs = 
       
       observeEvent(input$indVar, {
         values$indVar <- input$indVar
+      })
+      
+      observeEvent(input$rownamesAsIndVar, {
+        if (!input$rownamesAsIndVar) {
+          values$indVar <- input$indVar
+        } else {
+          values$indVar <- character(0)
+          updateSelectizeInput(session = session, "indVar", selected = character(0))
+        }
       })
       
       observeEvent(input$iter, {
@@ -124,12 +151,52 @@ modelSpecificationsServer <- function(id, dataMatrix, uploadedModelSpecInputs = 
         values$chains <- input$chains
       })
       
+      observeEvent(input$rndmSeed, {
+        values$rndmSeed <- input$rndmSeed
+      })
+      
+      observeEvent(input$fixedSeed, {
+        values$fixedSeed <- input$fixedSeed
+      })
+      
       reactive(values)
     })
   }
 
+#' Get Time Minimum 
+#' 
+#' @param mtrx (matrix) data matrix
+#' @param timeVars (character) column names of time variables
+#' @param default (numeric) default result
+#' 
+#' @return (numeric) minimal time
+#' @export
+getTimeMin <- function(mtrx, timeVars, default = 0) {
+  if (length(timeVars) == 0 || any(sapply(timeVars, function(x) x == ""))) {
+    return(default)
+  }
+  mtrx[, timeVars] %>% min()
+}
+
+#' Get Time Maximum 
+#' 
+#' @param mtrx (matrix) data matrix
+#' @param timeVars (character) column names of time variables
+#' @param default (numeric) default result
+#' 
+#' @return (numeric) maximal time
+#' @export
+getTimeMax <- function(mtrx, timeVars, default = 1) {
+  if (length(timeVars) == 0 || any(sapply(timeVars, function(x) x == ""))) {
+    return(default)
+  }
+  mtrx[, timeVars] %>% max()
+}
+
 defaultModelSpecValues <- function() {
   list(iter = 2000,
        burnin = 500,
-       chains = 4)
+       chains = 4,
+       rndmSeed = TRUE,
+       fixedSeed = 12345)
 }